refactor: 챔피언 메인 Model 객체에서 DTO 객체로 변경
